extend crud support to version objects via the api

+++ portality/models.py
@@ -252,6 +252,24 @@ class Version(LV, dao.VersionStoreDAO):
             self.data["version"][prop] = []
         self.data["version"][prop].append(val)
     
+    def patch_version(self, new_version, replace_all=True, keep_id=True):
+        # normalise the incoming version document
+        if "version" in new_version:
+            new_version = new_version.get("version")
+        
+        # remember the id
+        oid = None
+        if keep_id:
+            oid = self.id
+        
+        if replace_all:
+            self.data["version"] = new_version
+        else:
+            raise NotImplementedError()
+        
+        if keep_id:
+            self.id = oid
+    
     @property
     def id(self):
         return self.data.get("version", {}).get("id")
@@ -786,6 +804,53 @@ class LV_Index(object):
         return None
 
 class VersionIndex(LV_Index, dao.VersionIndexDAO):
+
+    @classmethod
+    def by_id(cls, version_id, cascade=True):
+        # get the original version
+        version = Version().get(version_id, links=True)
+        if version is None:
+            return
+        
+        # generate the version index
+        vi = VersionIndex.from_version(version)
+        vi.save()
+        
+        if not cascade:
+            return
+        
+        if version.singer is not None:
+            singer = Singer().get(version.singer, links=True)
+            if singer is not None:
+                si = SingerIndex.from_singer(singer)
+                si.save()
+        
+        if version.song is not None:
+            song = Song().get(version.song, links=True)
+            if song is not None:
+                si = SongIndex.from_song(song)
+                si.save()
+    
+    @classmethod
+    def delete_by_id(cls, version_id, cascade=True):
+        vi = VersionIndex.pull(version_id)
+        if vi is None:
+            return
+        
+        if vi.data.get("singer") is not None:
+            singer = Singer().get(vi.data.get("singer", {}).get("id"), links=True)
+            if singer is not None:
+                si = SingerIndex.from_singer(singer)
+                si.save()
+        
+        if vi.data.get("song") is not None:
+            song = Song().get(vi.data.get("song", {}).get("id"), links=True)
+            if song is not None:
+                si = SongIndex.from_song(song)
+                si.save()
+        
+        vi.delete()
+
     @classmethod
     def from_version(cls, version):
         # make a copy of the core version object and wrap the versionindex around it
@@ -859,39 +924,47 @@ class SingerIndex(LV_Index, dao.SingerIndexDAO):
     def by_id(cls, singer_id, cascade=True):
         # get the original singer
         singer = Singer().get(singer_id, links=True)
-        
-        if not cascade:
+        if singer is None:
             return
         
         # generate the singer index
         si = SingerIndex.from_singer(singer)
         si.save()
         
+        if not cascade:
+            return
+        
         # for each version regenerate the version index
         for v in singer.versions:
             version = Version().get(v, links=True)
-            vi = VersionIndex.from_version(version)
-            vi.save()
+            if version is not None:
+                vi = VersionIndex.from_version(version)
+                vi.save()
             
             # regenerate the song index
             if version.song is not None:
                 song = Song().get(version.song, links=True)
-                soi = SongIndex.from_song(song)
-                soi.save()
+                if song is not None:
+                    soi = SongIndex.from_song(song)
+                    soi.save()
     
     @classmethod
     def delete_by_id(cls, singer_id, cascade=True):
         si = SingerIndex.pull(singer_id)
+        if si is None:
+            return
         
         for v in si.data.get("versions"):
             version = Version().get(v.get("id"), links=True)
-            vi = VersionIndex.from_version(version)
-            vi.save()
+            if version is not None:
+                vi = VersionIndex.from_version(version)
+                vi.save()
             
             if version.get("song") is not None:
                 song = Song().get(version.get("song", {}).get("id"), links=True)
-                soi = SongIndex.from_song(song)
-                soi.save()
+                if song is not None:
+                    soi = SongIndex.from_song(song)
+                    soi.save()
             
         si.delete()
         
@@ -978,41 +1051,49 @@ class SongIndex(LV_Index, dao.SongIndexDAO):
     def by_id(cls, song_id, cascade=True):
         # get the original song
         song = Song().get(song_id, links=True)
-        
-        if not cascade:
+        if song is None:
             return
         
         # generate the singer index
         si = SongIndex.from_song(song)
         si.save()
         
+        if not cascade:
+            return
+        
         # for each version regenerate the version index
         for v in song.versions:
             version = Version().get(v, links=True)
-            vi = VersionIndex.from_version(version)
-            vi.save()
-            
-            # regenerate the singer index
-            if version.singer is not None:
-                singer = Singer().get(version.singer, links=True)
-                soi = SingerIndex.from_singer(singer)
-                soi.save()
+            if version is not None:
+                vi = VersionIndex.from_version(version)
+                vi.save()
+                
+                # regenerate the singer index
+                if version.singer is not None:
+                    singer = Singer().get(version.singer, links=True)
+                    if singer is not None:
+                        soi = SingerIndex.from_singer(singer)
+                        soi.save()
         
         # for each related song, regenerate its index
         for s in song.songs:
             relsong = Song().get(s, links=True)
-            ri = SongIndex.from_song(s)
-            ri.save()
+            if relsong is not None:
+                ri = SongIndex.from_song(s)
+                ri.save()
     
     @classmethod
     def delete_by_id(cls, song_id, cascade=True):
         si = SongIndex.pull(song_id)
+        if si is None:
+            return
         
         # refresh the indexes of related songs
         for rel in si.data.get("relations", []):
             rs = Song().get(rel.get("id"))
-            ri = SongIndex.from_song(s)
-            ri.save()
+            if rs is not None:
+                ri = SongIndex.from_song(s)
+                ri.save()
         
         # deleting a song deletes all the versions, so we need
         # to call delete by id on each version too
